# contracts/openapi/parkshield-api.v1.json
{
  "components": {
    "schemas": {
...
      "CoverageStatus": {
        "enum": [
          "VERIFIED_COVERAGE",
          "NO_VERIFIED_COVERAGE",
          "STALE_DATA",
          "LOCATION_PRECISION_INSUFFICIENT",
          "UNVERIFIABLE_SOURCE",
          "LOCATION_CONSENT_REQUIRED",
          "LOCATION_STALE"
        ],
        "title": "CoverageStatus",
        "type": "string"
      },
...
      "ErrorCode": {
        "description": "Initial public error-code catalog. Values are backward-compatible API surface.",
        "enum": [
          "INVALID_REQUEST",
          "AUTHENTICATION_REQUIRED",
          "AUTHENTICATION_FAILED",
          "SESSION_INVALID",
          "AUTHORIZATION_DENIED",
          "RESOURCE_NOT_FOUND",
          "CONFLICT",
          "VALIDATION_FAILED",
          "RATE_LIMITED",
          "UNSUPPORTED_MEDIA_TYPE",
          "PAYLOAD_TOO_LARGE",
          "SERVICE_UNAVAILABLE",
          "INTERNAL_ERROR",
          "NO_VERIFIED_COVERAGE",
          "STALE_DATA",
          "LOCATION_PRECISION_INSUFFICIENT",
          "DECISION_INDETERMINATE"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
      "ErrorDetail": {
        "additionalProperties": false,
        "description": "Allowlisted validation metadata; never includes submitted values or messages.",
        "properties": {
          "code": {
            "enum": [
              "MISSING_FIELD",
              "INVALID_FIELD"
            ],
            "title": "Code",
            "type": "string"
          },
          "field": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.\\[\\]-]+$",
            "title": "Field",
            "type": "string"
          }
        },
        "required": [
          "field",
          "code"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "description": "Versioned envelope returned for every handled API error.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "correlation_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Correlation Id",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ErrorDetail"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "version": {
            "const": "1",
            "default": "1",
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "correlation_id"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "LegacyParkingDecisionResponse": {
        "description": "Deprecated map response retained while Flutter remains a transition reference.",
        "properties": {
          "covered": {
            "title": "Covered",
            "type": "boolean"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "zone": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ParkingZoneResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "covered",
          "message"
        ],
        "title": "LegacyParkingDecisionResponse",
        "type": "object"
      },
...
      "ParkingDecisionEvidenceResponse": {
        "properties": {
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "import_batch_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Import Batch Id"
          },
          "observed_at": {
            "format": "date-time",
            "title": "Observed At",
            "type": "string"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "restriction_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Restriction Summary"
          },
          "source_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Id"
          },
          "zone_id": {
            "title": "Zone Id",
            "type": "string"
          },
          "zone_type": {
            "$ref": "#/components/schemas/ZoneType"
          }
        },
        "required": [
          "zone_id",
          "zone_type",
          "provenance",
          "confidence",
          "observed_at",
          "expires_at",
          "restriction_summary"
        ],
        "title": "ParkingDecisionEvidenceResponse",
        "type": "object"
      },
      "ParkingDecisionOutcome": {
        "enum": [
          "PARK",
          "CAUTION",
          "DO_NOT_PARK",
          "INDETERMINATE"
        ],
        "title": "ParkingDecisionOutcome",
        "type": "string"
      },
      "ParkingDecisionReasonCode": {
        "enum": [
          "OFFICIAL_RESTRICTION",
          "RESIDENT_PERMIT_REQUIRED",
          "TOWING_RISK",
          "NO_VERIFIED_COVERAGE",
          "STALE_DATA",
          "LOCATION_PRECISION_INSUFFICIENT",
          "DECISION_INDETERMINATE",
          "UNVERIFIABLE_SOURCE",
          "LOCATION_CONSENT_REQUIRED",
          "LOCATION_STALE"
        ],
        "title": "ParkingDecisionReasonCode",
        "type": "string"
      },
      "ParkingDecisionReasonResponse": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ParkingDecisionReasonCode"
          },
          "message": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ParkingDecisionReasonResponse",
        "type": "object"
      },
      "ParkingDecisionResponse": {
        "properties": {
          "coverage_status": {
            "$ref": "#/components/schemas/CoverageStatus"
          },
          "evaluated_at": {
            "format": "date-time",
            "title": "Evaluated At",
            "type": "string"
          },
          "evidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ParkingDecisionEvidenceResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "outcome": {
            "$ref": "#/components/schemas/ParkingDecisionOutcome"
          },
          "reasons": {
            "items": {
              "$ref": "#/components/schemas/ParkingDecisionReasonResponse"
            },
            "maxItems": 8,
            "minItems": 1,
            "title": "Reasons",
            "type": "array"
          }
        },
        "required": [
          "outcome",
          "coverage_status",
          "reasons",
          "evaluated_at"
        ],
        "title": "ParkingDecisionResponse",
        "type": "object"
      },
...
      "ParkingViewportResponse": {
        "properties": {
          "zones": {
            "items": {
              "$ref": "#/components/schemas/ParkingZoneResponse"
            },
            "title": "Zones",
            "type": "array"
          }
        },
        "required": [
          "zones"
        ],
        "title": "ParkingViewportResponse",
        "type": "object"
      },
      "ParkingZoneResponse": {
        "properties": {
          "average_towing_cost_cents": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Towing Cost Cents"
          },
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "geometry": {
            "additionalProperties": true,
            "title": "Geometry",
            "type": "object"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "observed_at": {
            "format": "date-time",
            "title": "Observed At",
            "type": "string"
          },
          "parking_score": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Parking Score",
            "type": "integer"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "restriction_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Restriction Summary"
          },
          "risk_level": {
            "$ref": "#/components/schemas/RiskLevel"
          },
          "towing_hotspot": {
            "title": "Towing Hotspot",
            "type": "boolean"
          },
          "zone_type": {
            "$ref": "#/components/schemas/ZoneType"
          }
        },
        "required": [
          "id",
          "name",
          "zone_type",
          "geometry",
          "parking_score",
          "risk_level",
          "provenance",
          "confidence",
          "restriction_summary",
          "towing_hotspot",
          "observed_at",
          "expires_at"
        ],
        "title": "ParkingZoneResponse",
        "type": "object"
      },
...
      "Provenance": {
        "enum": [
          "official",
          "community_verified",
          "ai_prediction",
          "estimated"
        ],
        "title": "Provenance",
        "type": "string"
      },
...
      "RiskLevel": {
        "enum": [
          "very_safe",
          "safe",
          "read_signs",
          "high_risk",
          "very_high_risk",
          "do_not_park"
        ],
        "title": "RiskLevel",
        "type": "string"
      },
...
      "ZoneType": {
        "enum": [
          "general",
          "resident_only",
          "private_property",
          "commercial",
          "towing_hotspot"
        ],
        "title": "ZoneType",
        "type": "string"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Parking intelligence with explicit source provenance.",
    "title": "ParkShield AI API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/parking/decision": {
      "get": {
        "deprecated": true,
        "operationId": "legacy_parking_decision_api_v1_parking_decision_get",
        "parameters": [
          {
            "in": "query",
            "name": "latitude",
            "required": true,
            "schema": {
              "maximum": 90,
              "minimum": -90,
              "title": "Latitude",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "longitude",
            "required": true,
            "schema": {
              "maximum": 180,
              "minimum": -180,
              "title": "Longitude",
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyParkingDecisionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication is required."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Access is denied."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request conflicts with current state."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Payload is too large."
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unsupported media type."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too many requests."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected server error."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service is temporarily unavailable."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Legacy Parking Decision",
        "tags": [
          "parking-map"
        ]
      }
    },
    "/api/v1/parking/decision/evaluate": {
      "get": {
        "operationId": "evaluate_parking_decision_api_v1_parking_decision_evaluate_get",
        "parameters": [
          {
            "in": "query",
            "name": "latitude",
            "required": true,
            "schema": {
              "maximum": 90,
              "minimum": -90,
              "title": "Latitude",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "longitude",
            "required": true,
            "schema": {
              "maximum": 180,
              "minimum": -180,
              "title": "Longitude",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "accuracy_meters",
            "required": true,
            "schema": {
              "maximum": 10000,
              "minimum": 0,
              "title": "Accuracy Meters",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "located_at",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "Located At",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "location_consent",
            "required": true,
            "schema": {
              "title": "Location Consent",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "has_resident_permit",
            "required": false,
            "schema": {
              "default": false,
              "title": "Has Resident Permit",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParkingDecisionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication is required."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Access is denied."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request conflicts with current state."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Payload is too large."
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unsupported media type."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too many requests."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected server error."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service is temporarily unavailable."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Evaluate Parking Decision",
        "tags": [
          "parking-map"
        ]
      }
    },
    "/api/v1/parking/zones": {
      "get": {
        "operationId": "viewport_api_v1_parking_zones_get",
        "parameters": [
          {
            "in": "query",
            "name": "west",
            "required": true,
            "schema": {
              "maximum": 180,
              "minimum": -180,
              "title": "West",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "south",
            "required": true,
            "schema": {
              "maximum": 90,
              "minimum": -90,
              "title": "South",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "east",
            "required": true,
            "schema": {
              "maximum": 180,
              "minimum": -180,
              "title": "East",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "north",
            "required": true,
            "schema": {
              "maximum": 90,
              "minimum": -90,
              "title": "North",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 500,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParkingViewportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication is required."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Access is denied."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request conflicts with current state."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Payload is too large."
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unsupported media type."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too many requests."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected server error."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service is temporarily unavailable."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Viewport",
        "tags": [
          "parking-map"
        ]
      }
    },
...
  }
}